set up scss global variables in vite config, create a global scss module, create a login scss module, create scss variables and function for easily accessing a color pallette
