+++ package.json
@@ -13,6 +13,7 @@
     "docker-down": "docker-compose -f docker-compose.yml down",
     "docker-dev-up": "docker-compose -f docker-compose-dev.yml up",
     "docker-dev-down": "docker-compose -f docker-compose-dev.yml down",
+    "docker-dev-build": "docker build -t docketeerxi/docketeer-dev .",
     "build": "vite build",
     "test": "jest --verbose",
     "lint": "eslint .",
