+++ src/components/Login/Login.module.scss
@@ -1,0 +1,31 @@
+// @use '../main.scss';
+
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 70vw;
+  height: 70vh;
+  margin: 0 15vw;
+}
+
+.input {
+  width: 100%;
+  height: 50px;
+  border: none;
+  border-radius: 5px;
+  padding: 0 1rem;
+  margin-bottom: 1rem;
+  font-size: 1.2rem;
+  color: $base-color;
+  background-color: #fff;
+}
+
+.formHolder {
+  display: flex;
+  flex-direction: column;
+}
+
+.logo {
+}
