+++ src/components/global.module.scss
@@ -1,0 +1,33 @@
+.button1 {
+  font-size: 1.5rem;
+  color: color(white, tan);
+  background-color: color(blue, dark);
+  border: none;
+  border-radius: 0.5rem;
+  padding: 1rem 2rem;
+  margin: 1rem 0;
+  cursor: pointer;
+  transition: all 0.3s ease-in-out;
+  &:hover {
+    background-color: #7a0d43;
+    color: #fff;
+    transform: scale(1.1);
+  }
+}
+
+.button2 {
+  font-size: 1.5rem;
+  color: color(blue, light);
+  background-color: transparent;
+  border: 2px solid color(blue, light);
+  border-radius: 0.5rem;
+  padding: 1rem 2rem;
+  margin: 1rem 0;
+  cursor: pointer;
+  transition: all 0.3s ease-in-out;
+  &:hover {
+    background-color: color(blue, dark);
+    color: #fff;
+    transform: scale(1.1);
+  }
+}
