+++ src/renderer/index.html
@@ -12,7 +12,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <div>WORK MF</div>
     <script type="module" src="./index.tsx"></script>
   </body>
 </html>
